Refactor: convert 'reciperadar.api.products._product_stream' from a generator method into a simple method with a single return value
Spoiler: this does, perhaps seemingly strangely, allow the flask-sqlalchemy application context lookup via werkzeug to succeed, providing a possible resolution for #65.  It's far from ideal though, because it negates the reason that we stream results in the response in the first place: the goal is to emit the results as they're received from the query without having to build up a complete result-set in the request's memory.  So this fix won't be acceptable to merge, but it's here for demonstration and confirmation purposes.